Force all calls to project_rect to include a bounding rect.
This greatly improves the accuracy of the clipper code when
required to calculate the world rect for a primitive that
crosses the near plane.

This patch by itself doesn't fix anything specific, but when
combined with the fixes that are in progress in the
plane-split crate, it will allow correct world rect calculations
for primitives that cross the near plane in various edge cases.