Add requested/actual composite mode to Picture primitive.
Previously, the composite_mode for a Picture was set during
display list flattening, and then not changed.

In the near future, we want to be able to decide during frame
build what the composite mode of a picture is. For example, if
we encounter a picture with a perspective transform, we may
decide to rasterize it in local space. Equally, if we encounter
a picture in a 3d rendering context, but detect there is no
perspective present and no plane splitting needed, we may
choose to skip drawing this to a surface and draw directly into
the parent.

This patch doesn't contain any functional changes - it just
introduces the concept of a requested and actual composite
mode, and updates surrounding code for this.