Fix order of compositing operations with a single stacking context.
A rare case is that a stacking context has both a mix-blend-mode
and a filter list on it.

The CSS compositing specification says that filters should be
applied first, and then any compositing.

Fixing this will also simplify some upcoming optimizations that
make mix-blend-mode much faster in commonly used scenarios.