Ensure the renderer always gets notified of transactions win whcih a frame is requested.
Gecko currently depends on transactions that request a frame to go all the way to the renderer and send back notifications that the frame is rendered. If such a notification gets lost, Gecko might wait forever that the frame get rendered before sending another frame request which means nothing ever gets rendered again.
This ensures we forward the information to the renderer even if for some reason we skipped building the frame, which tends to happen at stratup on Windows.