Refactor flattener for mix-blend-mode optimizations.
This patch doesn't contain any user visible changes. It does
change the display list flattener to append pictures to their
parent picture during pop_stacking_context instead of during
push_stacking_context.

This not only makes the code clearer, it also makes it possible
to introduce some optimizations to mix-blend-mode in the
future. Specifically, we want to detect mix blend mode, and
insert a new picture node in the tree which re-parents the
existing stacking context, in some cases.