Avoid allocating clip tasks for clip rects in same coordinate system.
If a clip chain only ends up with clip rects that are from the
same coordinate system, there is no need to allocate a clip mask,
as the clip can be implemented via the local clip rect in the
primitive vertex shader.

This is an important optimization for reducing clip task
allocations on common pages (although they won't actually be
drawn into the clip mask anyway).