Tidy up the scene flattening interface.
Now that display list flattening is always done by the scene
building thread, there is never a valid old frame builder
that can be recycled. Update the interface to make this
clearer.

Also slightly tidy up how picture ids are generated. This is
a temporary hack only - they won't be needed once we are
caching pictures based on their content.

This also removes the last current use or recycle_vec and
associated methods. I've removed them for now to avoid
warnings, but hopefully in the future we can find a way
to use these again in some places to avoid so many allocations
per frame!