Use raster roots for text shadows.
For shadows, establish a new rasterization root. This ensures
that text shadows are rasterized in local space, which in
turn makes them suitable to be cached between frames and
display lists when the content doesn't change.

The actual caching will be done as a follow up.

Also change the API from GlyphRasterSpace to RasterSpace, since
in the future we'll use it to signify that the entire stacking
context should be rasterized in local space, not just the glyphs.