Collect clip nodes for any surface, not just raster roots.
The existing code collects ancestor clips whenever we encounter
a rasterization root. This patch changes the behavior to
collect ancestor clips for any picture that has a surface.

This is a subtle but important step towards picture caching. It
ensures that any clips that are positioned by ancestors of the
surface positioning node are collected and applied when the
surface is composited into the parent, rather than on the
items in the surface itself. (The CSS specifications ensure
that this will give the equivalent result).

What this means is that a picture can be cached and re-used
on a subsequent frame, even if the clips in ancestor nodes
have moved relative to the positioning of the cached picture,
since those clips are only used during composition and do
not affect the contents of the picture itself.