Reduce number of calls to build clip chain instances.
A clip chain instance is built for each primitive. Then, during
segmentation we build a clip chain instance for each segment.

However, if the segmenting process only ends up with one segment
(the common case) we are effectively duplicating work by building
a clip chain instance which will be the same as the primitive
clip chain instance.

Detect this case and just use the primitive clip chain instance
if only one segment was produced.