Stop reinitializing freed texture cache textures.
We also drop the vec scheme and just use a hash-map. None of the operations
here are hot enough that the difference between a HashMap and a Vec
should matter, and using a HashMap simplifies things. Glenn says that
hashmap lookups were once hot in profiles, but that was before
FastHashMap and not specifically these lookups.

Differential Revision: https://phabricator.services.mozilla.com/D8164