Build border segments during scene building.
This patch changes border primitives to be able to
build their segments during scene building, rather
than frame building.

The main complexity was that we previously had to
do this during frame building to calculate the
device-space texel offsets for the segments. Now,
we supply this information as a local space stretch
rect for segment images. This allows the segments
for a border (and all primitives) to be treated
as immutable after first construction.

This is a CPU optimization (more so once the border
primitives are interned between display lists), but
most importantly it means that all segments can
now be built during scene building.

This in turn will allow a lot of the general segment
building and local space clips to be handled during
scene building and primitive interning, which is
both a performance win and makes possible some of
the future picture caching changes.