Get rid of the resolved_transform member variable.
This member variable is written to immediately before being used,
in SpatialNode::update_transform, and so can just be turned into
a local variable. The only other place that uses this is just a debug
printout. Finally, since the member variable can be removed, we can
also remove code that exists solely to populate this member.