Retain texture cache entries if fewer than 32
Noticed this when looking at how the caching interacts with
documents. Correct me if I'm wrong but this looks unintended - at
the very least it doesn't seem to match the comment.

Additionally, this moves the sorting inside the > 32 clause, since
there's no point in sorting if we're going to retain them all.