Move Primitive methods to PrimitiveInstance.
This is a small change towards reducing implementation details
on the Primitive struct. Eventually, Primitive will be removed
and we'll just have PrimitiveInstance and PrimitiveTemplate.

The PrimitiveTemplate struct will be the interned and shared
data for primitives that are the same across display lists.

The PrimitiveInstance struct currently contains information
calculated in prepare_prims and then used by batching. Once
we move the prepare prims code from the first (picture tree
pass) into the batching pass, most of the storage in the
PrimitiveInstance can be removed and used immediately.