Introduce picture traversal pass.
This introduces a new pass during frame building. It's a very
quick pass, since we only traverse the picture tree. This pass
must not access individual primitives.

In the future, the prepare_prims pass will be split into two
parts. Some of the work done in that pass will be moved to this
picture traversal pass (e.g. plane splitting), while the rest
(e.g GPU cache updates) will be moved to the batching pass.

The outcome of the first pass will determine whether a picture
has a valid cached surface, and thus decide whether the second
primitive preparation and batching pass is required to build
a new surface cache.