Refactor chase tracing to work with primitive instance IDs.
In the process of interning primitives, the PrimitiveIndex
will become less useful, eventually being removed since a
primitive instance will contain just a reference to an interned
primitive template.

Having the chase ID based on the PrimitiveIndex is complicating
some of that refactoring, so this patch switches it to create
a unique ID per primitive instance, which is then used by the
chase tracing code. The primitive ID only exists when the
tracing code is compiled in.