Apply opacity bindings in vertex shader, rather than CPU.
Previously, opacity bindings were calculated on the CPU and
stored in the GPU cache as part of the primitive color. This
required the GPU cache location to be invalidated when the
opacity binding changed.

This complicates primitive interning, where we want to be able
to share the GPU cache data for interned primitives, and not
need to update it during frame building.

Instead, provide the opacity for rectangles and images as part
of the primitive instance header, and apply in the vertex
shader.

This will simplify porting rectangles and images to use
primitive interning.