Add new notification checkpoint FrameTexturesUpdated.
FrameTexturesUpdated notifications should be issued when there are
either no pending texture updates, or all of the pending texture updates
have been applied. This is useful for listeners who would like to
perform an action after updating a resource. In particular this is
useful for Gecko to learn as soon as possible when it can release a
surface after it has swapped the external image ID of an image key such
that it points to a new buffer. Currently we rely upon FrameBuilt for
this, but future changes for recycling previously displayed surfaces for
animated images are more timing sensitive.