Use picture caching for text shadows.
This patch allows text shadows to be cached and persisted both
between frames and new display lists, in the texture cache.

Previously, text shadows could be cached between frames, but would
be flushed and redrawn when a new display list was processed.

The primitive and clip interning system is used to determine when
the contents of a cached picture is valid.

For now, this is only applied in very limited scenarios, in order
to ensure the core functionality is correct (specifically, blur
filters that contain glyphs and line decorations, which covers
most text shadows). In the future, this will be expanded to cover
(a) more primitive types (b) more surface / filter effects and
(c) at a finer granularity to allow partial picture caching (tiles).

As part of this change, we can remove all references to the global
scene id, as this is no longer required.

This brings the GPU time in the cloud ad [1] from 43 ms down to
8 ms for me. This is still higher than we want, due to the very high
number of cached text shadows that get composited. This should be
resolved as we support nested picture caching.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1436193