Optimize segment builder no-ops.
If we don't end up adding any clip regions to the segment builder,
then there is no point in going through the process of creating
events, sorting, and emitting segments.

Instead, detect this case and early out with a single segment.

This takes the backend time on the displaylist_mutate test on
my machine from 10.8ms to 8.2ms, which is quite a significant
improvement.