clip: Downscale box shadows earlier.
This fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1505953, and all shadows
with a large blur radius in general.

The key here is to not downscale the original allocation rect and the primitive
rect, which is what we use for stretching, but downscale the actual render task
and the filters.

This is a regression from #3230