Use glCopyImageSubData where available to copy data between texture arrays
Falling back to the existing glBlitFramebuffer loop. This avoids a
driver bug with texture arrays and glBlitFramebuffer on Adreno 4xx,
5xx, and 6xx devices, where the blit was failing and the texture cache
became corrupted.

Adreno 3xx seems to be affected by a similar bug, but does not support
glCopyImageSubData, so this can not be used there. Once the exact
nature of that bug has been established and a workaround identified,
we should re-evaluate whether this change is still desirable or if the
code paths can be consolidated.