Add PrimitiveScratchBuffer implementation.
This splits some information from the PrimitiveStore into the new
scratch buffer type.

The scratch buffer is specifically for storing information that
is only read / written during frame building, as opposed to other
items in the primitive store which is created during scene building.

Since the scratch buffer information is not moved between threads,
it's easy to retain it between both frame builds and also when a
new display list is sent by the scene builder thread.

When a new scene arrives, the recycle method allows the storage
of the scratch buffer to possibly be shrunk, if it is now
consuming much more memory than was used on the previous frame.

NOTE: This initial commit only uses the scratch buffer for clip
      mask instances, but there will be more fields moved to make
      use of the retained scratch buffer soon.