Port borders to be interned primitives.
This is mostly a straightforward conversion of borders
to make use of primitive interning.

It also adds support during batching and clip task creation for
interned primitives that make use of segments. This is needed
for porting the rest of the primitives, and should make the
conversion of the other primitive types simple.