Allow adding extra cargo flags via the CARGOFLAGS env var
This allows Firefox CI to pass the --frozen flag to all the cargo
commands, which will ensure that we are pulling the vendored
dependencies correctly and pulling from crates.io during test runs.

It is also an added convenience for developers if they want to pass
extra cargo debugging flags to all the cargo invocations.