Add RectangleKey for hashing float rects.
The Au <-> f32 conversion is showing up in a lot of talos profiles.

Since layout in both Gecko and Servo already operate on Au, the
values are already quantized, and so it's reasonable to just
directly hash the floats and avoid the conversions.

This patch changes the main rectangles in PrimitiveKey to be
directly hashed. We can incrementally convert other fields in
the PrimitiveKeyKind to also hash the data types directly.

On Talos, this takes the dl_mutate time from ~4780 down to ~4600
which is a decent win.