Store clip node templates in local space.
This change makes clip item keys independent of the local space
origin of the node. This ensures that even when Gecko sends a
new display list with clip nodes that have different scroll offsets
the interning logic results in the same uid for the clip node,
avoiding unnecessary invalidations of cached picture tiles.