Fix a BlobImageRasterizer race condition.
We can only make a blob image rasterizer the render backend thread's current rasterizer after it went through the scene builder thread, otherwise the rastirezer might race ahead of other rasterizers that are on going through the scene builder thread.

We also need to make sure that if there is a blob image update, we don't skip the scene builder thread.