Move primitive origin from template data to the instance.
This allows sharing of primitive templates where the only difference
between two primitives is the local space position.

This also means that interning works as expected for primitives
sent by Gecko that contain baked in scroll offsets.

This patch changes the primitive rect only - the next step is a
follow up patch that makes the clip rect, and a few other params
in PrimitiveTemplateKind to also be in a true local space.

Also normalize the clip rect to the primitive origin.